make DashboardNavbar a pure component and use id for list keys
